@@ -24,7 +24,8 @@
     "forum": "http://discourse.roots.io/"
   },
   "require": {
-    "php": ">=5.3.0",
-    "composer/installers": "~1.0"
+    "php": ">=5.4.0",
+    "composer/installers": "~1.0",
+    "jwage/purl": "^0.0.6"
   }
 }
